% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanText.R
\name{cleanText}
\alias{cleanText}
\title{A basic function to process text}
\usage{
cleanText(TEXT, lower = NULL, removeNumbers = NULL,
  removePunctuation = NULL, removeStopWords = NULL,
  removeSingleLetters = NULL, verbose = T)
}
\arguments{
\item{TEXT}{a string vector}

\item{lower}{convert all characters in TEXT to lower-case?  Defaults to T}

\item{removeNumbers}{remove all numbers from TEXT?  Defaults to T}

\item{removePunctuation}{remove all PUNCTUATION from TEXT?  Defaults to T}

\item{removeStopWords}{remove all stop words from TEXT?  Defaults to T.  Default list of stop words is the "en" list in the tm package.}

\item{removeSingleLetters}{remove all single letters from TEXT?  Defaults to T.}

\item{verbose}{a logical vector for output to console.  Defaults to F.}
}
\description{
This function allows you to count flex grams given a string vector
}
\examples{
cleanText(remarks)


}
\keyword{processing}
\keyword{text}
