% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexGramCount.R
\name{flexGramCount}
\alias{flexGramCount}
\title{Calucluate the most frequent flex-grams}
\usage{
flexGramCount(TEXT, maxN = NULL, minN = NULL, minCount = NULL,
  verbose = NULL)
}
\arguments{
\item{TEXT}{a string vector}

\item{maxN}{the maximum number of unigrams allowed in a flexgram.  Defaults to 5.}

\item{minN}{the minimum number of unigrams allowed in a flexgram.  Defaults to 2.}

\item{minCount}{the minimum number of counts required for each flex-gram.  Defaults to min(floor(0.01*length(TEXT)) , 2).}

\item{verbose}{a logical vector for output to console.  Defaults to F.}
}
\description{
This function allows you to count flex grams given a string vector
}
\examples{
flexGramCount(remarks)



}
\keyword{flex-grams}
