% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedonicWithText.R
\name{hedonicWithText}
\alias{hedonicWithText}
\title{Estimate a hedonic model}
\usage{
hedonicWithText(Y, X = NULL, M = NULL)
}
\arguments{
\item{Y}{the dependent variable.}

\item{X}{a matrix or vector of attributes for the house.  This is not required.}

\item{M}{a matrix for the tokens.  This is required.}
}
\description{
This function allows you to estimate a hedonic model using an attribute matrix and a token matrix.
}
\examples{
textFit <- hedonicWithText(logprice , attributeMat , tokenMat)


}
\keyword{LASSO}
\keyword{hedonic}
\keyword{model,}
