% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenMatrix.R
\name{tokenMatrix}
\alias{tokenMatrix}
\title{A basic function to process text}
\usage{
tokenMatrix(TEXT, TOKENLIST, verbose = NULL)
}
\arguments{
\item{TEXT}{a string vector.}

\item{TOKENLIST}{a string vector with the tokens to look for in TEXT.}

\item{verbose}{a logical vector for output to console.  Defaults to F.}
}
\description{
This function allows you to count flex grams given a string vector
}
\examples{
M <- tokenMatrix(remarks , realEstateTokens)


}
\keyword{a}
\keyword{each}
\keyword{for}
\keyword{indicator}
\keyword{make}
\keyword{matrix}
\keyword{of}
\keyword{token}
\keyword{variables}
