% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenTable.R
\name{tokenTable}
\alias{tokenTable}
\title{A basic function to process text}
\usage{
tokenTable(B, ROUND = NULL)
}
\arguments{
\item{B}{a coefficient vector from a glmnet fit.}

\item{ROUND}{round the coefficeints?  Defaults to T}
}
\description{
This function allows you to count flex grams given a string vector
}
\examples{
b <- coef(unigramFit , s='lambda.min')
tokenDataFrame <- tokenTable(b)


}
\keyword{a}
\keyword{and}
\keyword{loadings}
\keyword{make}
\keyword{table}
\keyword{tokens}
\keyword{with}
