% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexGramCount.R
\name{flexGramCount}
\alias{flexGramCount}
\title{Calucluate the most frequent flex-grams}
\usage{
flexGramCount(TEXT, MAXN = NULL, MINN = NULL, MINCOUNT = NULL,
  VERBOSE = NULL)
}
\arguments{
\item{TEXT}{a string vector}

\item{MAXN}{the maximum number of unigrams allowed in a flexgram.  Defaults to 5}

\item{MINN}{the minimum number of unigrams allowed in a flexgram.  Defaults to 2}

\item{MINCOUNT}{the minimum number of counts required for each flex-gram.  Defaults to min{floor(0.01*length(TEXT)) , 2}}

\item{VERBOSE}{a local vector for output to console}
}
\description{
This function allows you to count flex grams given a string vector
}
\examples{
flexGramCount(remarks)
}
\keyword{flex-grams}
